backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-dotenv
python-jose
passlib
bcrypt>=4.0.0
jwt
python-multipart
google-auth
requests
# Database 
pymongo
psycopg2-binary

#Agents
groq